Create HTML (TABLE AND TAGS)
This project creates a visually appealing web page suitable for displaying a class timetable. It blends aesthetic design elements with functional content, making it an engaging and informative web page for students or anyone interested in the displayed timetable. The use of custom fonts, background video, and organized table structure contributes to the project's overall visual appeal.